CN/R, CN-Half added, Perf + 30% on all algos

+++ src/slow-hash-portable.c
@@ -98,10 +98,10 @@ static void xor64(uint8_t *a, const uint64_t b) //Perf +30%
   *(uint64_t *)a ^= b;
 }
 
-void cn_slow_hash(const void *data, size_t length, char *hash, int light, int variant, int prehashed, uint32_t page_size, uint32_t scratchpad, uint32_t iterations, int height)
+void cn_slow_hash(const void *data, size_t length, char *hash, int light, int variant, int prehashed, uint64_t page_size, uint64_t scratchpad, uint64_t iterations, int height)
 {
-    uint32_t init_rounds = (scratchpad / INIT_SIZE_BYTE);
-    uint32_t aes_rounds = (iterations / 2);
+    uint64_t init_rounds = (scratchpad / INIT_SIZE_BYTE);
+    uint64_t aes_rounds = (iterations / 2);
     size_t lightFlag = (light ? 2: 1);
 
     uint8_t text[INIT_SIZE_BYTE];
@@ -164,8 +164,8 @@ void cn_slow_hash(const void *data, size_t length, char *hash, int light, int va
 
     for(i = 0; i < aes_rounds; i++)
     {
-    #define MASK(div) ((uint32_t)(((page_size / AES_BLOCK_SIZE) / (div) - 1) << 4))
-    #define state_index(x,div) ((*(uint32_t *) x) & MASK(div))
+    #define MASK(div) ((uint64_t)(((page_size / AES_BLOCK_SIZE) / (div) - 1) << 4))
+    #define state_index(x,div) ((*(uint64_t *) x) & MASK(div))
 
       // Iteration 1
       j = state_index(a,lightFlag);
